django_email-main/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
djangorestframework==3.12.4
greenlet==1.1.2
idna==3.3
mssql-django==1.0
Pillow==8.4.0
psycopg2==2.9.1
pymssql==2.2.2
pyodbc==4.0.32
pytz==2021.3
requests==2.26.0
SQLAlchemy==1.4.25
sqlparse==0.4.2
urllib3==1.26.7
